Add pytest configuration for break on exception
Pytest nicely catches raised exceptions during tests and reports them
at the end. This is very useful except when debugging tests in vscode
because it prevents the debugger from pausing when there is an exception
and displaying the call stack.

This adds hooks for pytest to re-raise the caught exceptions when an
environment variable is set, and then sets that environment variable
in the `Debug Unit Test` launch configuration.