ci: Remove check workflow and filter on branch name (#253)
Removes the check workflow and using it to deduplicate PRs and pushes:
CI should be run on all PRs and also on `main` branch: pushes to PRs are
already picked up by the pull_request filtering: "By default, a workflow
only runs when a pull_request event's activity type is opened,
synchronize, or reopened.", the definition of `synchronize` can be found
in the documentation: "A pull request's head branch was updated. For
example, the head branch was updated from the base branch or new commits
were pushed to the head branch."

E: TC believes someone at some point wanted CI to run on branches that
do not have open PRs, because when a PR is opened he gets notified
whether it is draft or not. This does require that now a draft PR is
opened for CI to run, but does anyone have a use case for running CI
against a branch without even a draft PR?