fix: Handle pinned sha versions of Python in the install_requirements action (#268)
When PYTHON_VERSION is pinned to a specific sha (as done by Renovate
best-practices), install_requirements fails as the container tag is not
a valid version of python when including the sha. This trims the
trailing part of the tag to just the [0-9\.]+ part which should be the
valid python version.